During polishing smoke particle system

+++ PXCore/Particle/Particle.h
@@ -11,7 +11,7 @@ namespace Core::Particle {
 		void Draw(sf::Image& image);
 		bool ToDelete(const sf::Vector2u& image_size)const;
 	protected:
-		sf::Vector2i _position;
+		sf::Vector2f _position;
 		sf::Vector2f _velocity;
 		sf::Color _color;
 		float _time{ .0f };
